AI-Powered-Text-Summarizer-And-Voice-Translator-Web-app/ai-voice translator/text_summarizer_project/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
transformers==4.35.0
torch==2.1.0
PyMuPDF==1.23.8
python-docx==0.8.11
gTTS==2.4.0
reportlab==4.0.7
sentencepiece==0.1.99
protobuf==4.24.4
